requirements.txt
# 依赖 zeval-ai 核心框架
-e ..

# 注意：代码中已包含 torch.xpu 的 monkey patch，兼容 PyTorch 2.2.2

# FastAPI 相关
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6  # 文件上传支持

# 工具库
python-dotenv>=1.0.0
rich>=13.0.0
schedule>=1.2.0
httpx>=0.27.0  # RAG API 调用

# 数据处理
openpyxl>=3.1.0
PyPDF2>=3.0.0

# 未来需要的依赖（暂时注释）
# fastapi>=0.100.0
# uvicorn>=0.20.0
# sqlalchemy>=2.0.0
